PC104 interrupt handling changes, adjust MSC0 for dm9000
In titan_gpio_pc104_handler(), if pending is initially 0, set
it to the unmasked interrupts. This seems to help avoid
the problem of PC104 interrupts ceasing to be serviced, which
happens especially under system load. When this happens,
the PC104/ISA IRQ line, in this case IRQ=3, is high, but the handler
is never called again, unless the IRQ signal is shorted low,
or the device (tty port) is re-opened.

Using #ifdefs, revert to using handle_simple_irq() instead of
handle_simple_irq_til_done() for PC104  interrupts, in order to
test if it is needed.

Change MSC0_RDF1 from 5 to 7, which seems to cure the "status check fail"
problem with the dm9000.