Updates for 3.16 on Eurotech viper
Generally working, needs testing. PC104 has not been tested.
Had to re-work the iomap in arch/arm/mach-pxa/include/mach/io.h
for PCMCIA/ATA accesses to work on a Viper. Can't fiddle with
addresses of the PCCARD slot 0 in the same was as is done for
the PC104 ioports.

TODO: test MSC1 settings to see if the USB isp116x delay can be removed.