Updates to PC104 interrupt handling
Currently PC104 interrupts are handled with a threaded handler,
with TRY_HEADED_HANDER, defined.  Thinks work either way, not
sure which is best.

Rather mysteriously (to me at least) had to pass the irq
as an unsigned int to generic_handle_irq(uirq). Otherwise
got the "nobody cared" message on IRQ3.

Also added a comment that the Viper won't boot if the
i2c_bus_device is enabled in viper_devs.