Added compile time macro, TRY_THREADED_HANDLER
In trying to resolve a dropped interrupt issue with
Brandywine PC104SG cards both threaded and non-threaded
handlers were tried. The problem was resolved by modifying the
driver, and a thread or non-threaded handler had no effect.
But we'll leave the  macro in.

Revert back to handle_simple_irq() instead of the
handle_simple_irq_til_done() hack handler for the PC104 interrupts.