Fix for V2 PC104 IRQ bug, updates to USB2 GPIO
In V2 Titans, PC104 IRQs 10 and 11 are in bits 6 and 7 of
TITAN_LO_IRQ_STATUS instead of bits 5 and 6.  IRQ 12 is
not detectable.

Remove initialization of GPIO119_USBH2_PWR, GPIO120_USBH2_PEN
which are not connected to USB port 2 on a Titan.
GPIO114 and GPIO22 are used instead. For GPIO22, do
    MFP_CFG_OUT(GPIO22, AF0, DRIVE_LOW)
Previously it was
    MFP_CFG_OUT(GPIO22, AF0, DRIVE_HIGH)
The last parameter to MFP_CFG_OUT is supposed to be the low power state.
According to the Titan manual, the sleep state is 0, low.