RTC on ds1338 is now working
Testing found the right values for the I2C "is_open_drain" parameters
for SDA and SCL on GPIO83 and 84. With these values the bus can be
driven at 100kHz, with a udelay=5, and a minimum timeout of 1 jiffie.

On Viper V1 the pxa i2c bus is not enabled. This needs testing on
V2 vipers.

A wrong I2C config can cause a kernel hang when it tries to update
the RTC.